+++ connector/CHANGES.rst
@@ -1,6 +1,13 @@
 Changelog
 ---------
 
+2.1.1 (2014-02-06)
+~~~~~~~~~~~~~~~~~~
+
+* A user can be blocked because he has no access to the model queue.job when a
+  job has been delayed. The creation of a job is low level and should not be
+  restrained by the accesses of the user. (lp:1276182)
+
 2.1.0 (2014-01-15 - warning: breaks compatibility)
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~